:truck: refactor: move bcript-adapter files to a folder
