+++ globalConfig.json
@@ -1,1 +1,1 @@
-{"mongoUri":"mongodb://127.0.0.1:50612/jest?","mongoDBName":"jest"}
+{"mongoUri":"mongodb://127.0.0.1:50349/jest?","mongoDBName":"jest"}
