+++ src/infra/criptography/bcrypt-adapter/bcrypt-adapter.ts
@@ -1,6 +1,6 @@
 import bcrypt from 'bcrypt'
-import { Hasher } from '../../data/protocols/criptography/hashers'
-import { HashComparer } from '../../data/protocols/criptography/hash-comparer'
+import { Hasher } from '../../../data/protocols/criptography/hashers'
+import { HashComparer } from '../../../data/protocols/criptography/hash-comparer'
 
 export class BcryptAdapter implements Hasher, HashComparer {
   private readonly salt: number
